fix: Fix unit null rank
